requirements.txt
streamlit
sqlalchemy
psycopg2-binary
pydantic

streamlit>=1.33.0
sqlalchemy>=2.0.29
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
passlib[bcrypt]>=1.7.4
pandas>=2.2.2
